backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=1.4.0
alembic>=1.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
psycopg2-binary>=2.9.0
requests>=2.26.0
aiohttp>=3.8.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.0
pytest>=6.2.5
httpx>=0.19.0
tenacity>=8.0.0
celery>=5.1.0
redis>=4.0.0
gunicorn>=20.1.0
sentry-sdk>=1.5.0
geopy>=2.2.0
openai>=1.0.0
